refactor(plugins): 统一内置与第三方插件级元数据并精简注册体系
**🤔 背景与动机 (Why)**
- 插件管理页表格与详情走不同数据通路，内置条目信息双源不一致。
- 内置判定依赖否定式推断或组件白名单，绑定注册表与运行时列表的成对不变式，脆弱且易误判。
- 插件信息缓存体系（双列表与状态类型）零消费，与注册表互为冗余镜像。

**✨ 解决方案与影响 (What & Impact)**
- 元数据增加宿主管辖的类型枚举字段，内置与第三方分别由构造与加载强制标注，判定改为枚举判断。
- 列表、详情与命令行入口统一收敛到单一组装方法，优先级全为非负整数且排序口径统一为插件声明值。
- 退役插件信息缓存与内置双列表，注册表取消元数据缓存改直接迭代取源，详情按标识直查运行时。
- 前端统一行模型并补充运行状态展示，详情弹窗入口文案与交互细节一并修正。

+++ crates/plugin-sdk-rust/src/runtime.rs
@@ -377,7 +377,7 @@ async fn dispatch(
                     component_description: ds.component_description().to_string(),
                     component_type: ds.component_type(),
                     kind: ComponentKind::DataSource,
-                    priority: ds.priority() as i32,
+                    priority: ds.priority(),
                 });
             }
             for ex in &app.executors {
@@ -389,7 +389,7 @@ async fn dispatch(
                     kind: ComponentKind::ActionExecutor {
                         target_types: ex.supported_target_types(),
                     },
-                    priority: ex.priority() as i32,
+                    priority: ex.priority(),
                 });
             }
             Ok(serde_json::to_value(components).unwrap_or_default())
